+++ indra/statements.py
@@ -1,6 +1,4 @@
 import warnings
-import itertools
-from sets import ImmutableSet
 from pysb import *
 from pysb import ReactionPattern, ComplexPattern, ComponentDuplicateNameError
 
@@ -78,32 +76,6 @@ def site_name(stmt):
 
     return names
 
-def get_create_parameter(model, name, value, unique=True):
-    """Return parameter with given name, creating it if needed.
-
-    If unique is false and the parameter exists, the value is not changed; 
-    if it does not exist, it will be created. If unique is true then upon conflic 
-    a number is added to the end of the parameter name.  
-    """
-    parameter = model.parameters.get(name)
-
-    if not unique and parameter is not None:
-        return parameter
-    
-    if unique:
-        pnum = 1
-        while True:
-            pname = name + '_%d' % pnum
-            if model.parameters.get(pname) is None:
-                break
-            pnum += 1
-    else:
-        pname = name
-
-    parameter = Parameter(pname, value)
-    model.add_component(parameter)
-    return parameter
-
 def get_activating_mods(agent, agent_set):
     act_mods = agent_set[agent.name].activating_mods
     if not act_mods:
@@ -113,40 +85,6 @@ def get_activating_mods(agent, agent_set):
 def get_binding_site_name(name):
     binding_site = name.lower()
     return binding_site
-
-def get_complex_pattern(model, agent, agent_set, extra_fields=None):
-    '''
-    Constructs a PySB ComplexPattern from an Agent
-    '''
-    monomer = model.monomers[agent.name]
-    
-    pattern = {}
-
-    if extra_fields is not None:
-        for k,v in extra_fields.iteritems():
-            pattern[k] = v
-    if agent.bound_to:
-        # Here we make the assumption that the binding site
-        # is simply named after the binding partner
-        if agent.bound_neg:
-            pattern[get_binding_site_name(agent.bound_to)] = None
-        else:
-            pattern[get_binding_site_name(agent.bound_to)] = ANY
-    
-    # Add the pattern for the modifications of the agent
-    # TODO: This is specific to phosphorylation but we should be 
-    # able to support other types as well
-    for m, mp in zip(agent.mods, agent.mod_sites):
-        mod = abbrevs[m]
-        mod_pos = mp if mp is not None else ''
-        mod_site = ('%s%s' % (mod, mod_pos))
-        pattern[mod_site] = 'p'
-    
-    complex_pattern = monomer(**pattern)
-    return complex_pattern
-
-class UnknownPolicyException(Exception):
-    pass
 
 
 class Agent(object):
@@ -210,16 +148,6 @@ class Statement(object):
         else:
             return False
 
-    def monomers(self, agent_set, policies=None):
-        """Calls the appropriate monomers method based on policies."""
-        if policies is None or policies == 'one_step':
-            self.monomers_one_step(agent_set)
-        elif policies == 'interactions_only':
-            self.monomers_interactions_only(agent_set)
-        elif policies == 'two_step':
-            self.monomers_two_step(agent_set)
-        else:
-            raise UnknownPolicyException(policies)
 
     def assemble(self, model, agent_set, policies=None):
         """Calls the appropriate assemble method based on policies."""
@@ -990,177 +918,7 @@ class Complex(Statement):
                 return False
         return True
 
-    def monomers_interactions_only(self, agent_set):
-        return self.monomers_one_step(agent_set)
 
-    def assemble_interactions_only(self, model, agent_set):
-        return self.assemble_one_step(model, agent_set)
-
-    def monomers_one_step(self, agent_set):
-        """In this (very simple) implementation, proteins in a complex are
-        each given site names corresponding to each of the other members
-        of the complex (lower case). So the resulting complex can be
-        "fully connected" in that each member can be bound to 
-        all the others."""
-        for i, member in enumerate(self.members):
-            gene_mono = agent_set.get_create_base_agent(member)
-            # Add sites for agent modifications
-            # TODO: This assumes phosphorylation, but in principle
-            # it could be some other modification
-            for m, mp in zip(member.mods, member.mod_sites):
-                mod = abbrevs[m]
-                mod_pos = mp if mp is not None else ''
-                mod_site = ('%s%s' % (mod, mod_pos))
-                gene_mono.create_site(mod_site, ['u', 'p'])
-            
-            # Specify a binding site for each of the other complex members
-            # bp = abbreviation for "binding partner"
-            for j, bp in enumerate(self.members):
-                # The protein doesn't bind to itself!
-                if i == j:
-                    continue
-                gene_mono.create_site(get_binding_site_name(bp.name))
-    
-    def assemble_one_step(self, model, agent_set):
-        pairs = itertools.combinations(self.members, 2)
-        for pair in pairs:
-            agent1 = pair[0]
-            agent2 = pair[1]
-            param_name = agent1.name[0].lower() +\
-                         agent2.name[0].lower() + '_bind'
-            kf_bind = get_create_parameter(model, 'kf_' + param_name, 1e-6)
-            kr_bind = get_create_parameter(model, 'kr_' + param_name, 1e-6)
-
-            # Make a rule name
-            name_components = []
-            for m in pair:
-                if m.bound_to:
-                    if m.bound_neg:
-                        name_components.append(m.name + '_n' + m.bound_to)
-                    else:
-                        name_components.append(m.name + '_' + m.bound_to)
-                else:
-                    name_components.append(m.name)
-            
-            # Construct full patterns of each agent with conditions
-            rule_name = '_'.join(name_components) + '_bind'
-            agent1_pattern = get_complex_pattern(model, agent1, agent_set)
-            agent2_pattern = get_complex_pattern(model, agent2, agent_set)
-            agent1_bs = get_binding_site_name(agent2.name)
-            agent2_bs = get_binding_site_name(agent1.name)
-            r = Rule(rule_name, agent1_pattern(**{agent1_bs: None}) +\
-                                agent2_pattern(**{agent2_bs: None}) >>
-                                agent1_pattern(**{agent1_bs: 1}) %\
-                                agent2_pattern(**{agent2_bs: 1}), 
-                                kf_bind)
-            add_rule_to_model(model, r)
-
-            # In reverse reaction, assume that dissocition is unconditional
-            rule_name = '_'.join(name_components) + '_dissociate'
-            agent1_uncond = get_complex_pattern(model, Agent(agent1.name), agent_set)
-            agent2_uncond = get_complex_pattern(model, Agent(agent2.name), agent_set)
-            r = Rule(rule_name, agent1_uncond(**{agent1_bs: 1}) %\
-                                agent2_uncond(**{agent2_bs: 1}) >>
-                                agent1_uncond(**{agent1_bs: None}) +\
-                                agent2_uncond(**{agent2_bs: None}), 
-                                kr_bind)
-            add_rule_to_model(model, r)
-
-    def assemble_multi_way(self, model, agent_set):
-        # Get the rate parameter
-        abbr_name = ''.join([m.name[0].lower() for m in self.members])
-        kf_bind = get_create_parameter(model, 'kf_' + abbr_name + '_bind', 1e-6)
-        kr_bind = get_create_parameter(model, 'kr_' + abbr_name + '_bind', 1e-6)
-
-        # Make a rule name
-        name_components = []
-        for m in self.members:
-            if m.bound_to:
-                if m.bound_neg:
-                    name_components.append(m.name + '_n' + m.bound_to)
-                else:
-                    name_components.append(m.name + '_' + m.bound_to)
-            else:
-                name_components.append(m.name)
-        rule_name = '_'.join(name_components) + '_bind'
-        # Initialize the left and right-hand sides of the rule
-        lhs = ReactionPattern([])
-        rhs = ComplexPattern([], None)
-        # We need a unique bond index for each pair of proteins in the
-        # complex, resulting in n(n-1)/2 bond indices for a n-member complex.
-        # We keep track of the bond indices using the bond_indices dict,
-        # which maps each unique pair of members to a bond index.
-        bond_indices = {}
-        bond_counter = 1
-        for i, member in enumerate(self.members):
-            gene_name = member.name
-            mono = model.monomers[gene_name]
-            # Specify free and bound states for binding sites for each of
-            # the other complex members
-            # (bp = abbreviation for "binding partner")
-            left_site_dict = {}
-            right_site_dict = {}
-            for j, bp in enumerate(self.members):
-                bp_bs = get_binding_site_name(bp.name)
-                # The protein doesn't bind to itself!
-                if i == j:
-                    continue
-                # Check to see if we've already created a bond index for these
-                # two binding partners
-                bp_set = ImmutableSet([i, j])
-                if bp_set in bond_indices:
-                    bond_ix = bond_indices[bp_set]
-                # If we haven't see this pair of proteins yet, add a new bond
-                # index to the dict
-                else:
-                    bond_ix = bond_counter
-                    bond_indices[bp_set] = bond_ix
-                    bond_counter += 1
-                # Fill in the entries for the site dicts
-                left_site_dict[bp_bs] = None
-                right_site_dict[bp_bs] = bond_ix
-            
-            # Add the pattern for the modifications of the member
-            # TODO: This is specific to phosphorylation but we should be 
-            # able to support other types as well
-            for m, mp in zip(member.mods, member.mod_sites):
-                mod = abbrevs[m]
-                mod_pos = mp if mp is not None else ''
-                mod_site = ('%s%s' % (mod, mod_pos)) 
-                left_site_dict[mod_site] = 'p'
-                right_site_dict[mod_site] = 'p'
-
-            # Add the pattern for the member being bound
-            if member.bound_to:
-                bound_name = member.bound_to
-                bound_bs = get_binding_site_name(bound_name)
-                gene_bs = get_binding_site_name(gene_name)
-                if member.bound_neg:
-                    left_site_dict[bound_bs] = None
-                    right_site_dict[bound_bs] = None
-                    left_pattern = mono(**left_site_dict)
-                    right_pattern = mono(**right_site_dict)
-                else:
-                    bound = model.monomers[bound_name]
-                    left_site_dict[bound_bs] =\
-                        bond_counter
-                    right_site_dict[bound_bs] =\
-                        bond_counter
-                    left_pattern = mono(**left_site_dict) % \
-                                    bound(**{gene_bs:bond_counter})
-                    right_pattern = mono(**right_site_dict) % \
-                                    bound(**{gene_bs:bond_counter})
-                    bond_counter += 1 
-            else:
-                left_pattern = mono(**left_site_dict)
-                right_pattern = mono(**right_site_dict)
-            # Build up the left- and right-hand sides of the rule from
-            # monomer patterns with the appropriate site dicts
-            lhs = lhs + left_pattern
-            rhs = rhs % right_pattern
-        # Finally, create the rule and add it to the model
-        rule = Rule(rule_name, lhs <> rhs, kf_bind, kr_bind)
-        add_rule_to_model(model, rule)
 
     def __str__(self):
         return ("Complex(%s)" % [m.name for m in self.members])
